requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
pymongo==4.10.1
cohere==5.3.0
pinecone-client==5.0.1
openai==1.51.0
pypdf==5.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
